RavenDB-17492 : move AssertNotIncrementalTimeSeriesForLegacyReplication from CanContinueBatch into AssertNoLegacyReplicationViolation, add test
