+++ test/InterversionTests/ReplicationTests.cs
@@ -111,47 +111,7 @@ namespace InterversionTests
             }
         }
 
-        [Fact]
-        public async Task ShouldNotReplicateIncrementalTimeSeriesToOldServer2()
-        {
-            const string version = "5.2.3";
-            const string incrementalTsName = Constants.Headers.IncrementalTimeSeriesPrefix + "HeartRate";
-            const string docId = "users/1";
-            var baseline = DateTime.UtcNow;
-
-            using (var oldStore = await GetDocumentStoreAsync(version))
-            using (var store = GetDocumentStore())
-            {
-                using (var session = store.OpenAsyncSession())
-                {
-                    await session.StoreAsync(new User { Name = "ayende" }, docId);
-                    session.IncrementalTimeSeriesFor(docId, incrementalTsName)
-                        .Increment(baseline, 1);
-                    await session.SaveChangesAsync();
-                }
-
-                using (var session = store.OpenAsyncSession())
-                {
-                    var u = await session.LoadAsync<User>(docId);
-                    session.IncrementalTimeSeriesFor(u, incrementalTsName)
-                        .Increment(baseline, 1);
-
-                    u.Name = "oren";
-
-                    await session.SaveChangesAsync();
-                }
-
-                await SetupReplication(store, oldStore);
-
-                var replicationLoader = (await GetDocumentDatabaseInstanceFor(store)).ReplicationLoader;
-                Assert.NotEmpty(replicationLoader.OutgoingFailureInfo);
-                Assert.True(WaitForValue(() => replicationLoader.OutgoingFailureInfo.Any(ofi => ofi.Value.RetriesCount > 2), true));
-                Assert.True(replicationLoader.OutgoingFailureInfo.Any(ofi => ofi.Value.Errors.Any(x => x.GetType() == typeof(LegacyReplicationViolationException))));
-                Assert.True(replicationLoader.OutgoingFailureInfo.Any(ofi => ofi.Value.Errors.Select(x => x.Message).Any(x => x.Contains("IncrementalTimeSeries"))));
-            }
-        }
-
-        private static async Task<ModifyOngoingTaskResult> SetupReplication(IDocumentStore src, IDocumentStore dst)
+        protected static async Task<ModifyOngoingTaskResult> SetupReplication(IDocumentStore src, IDocumentStore dst)
         {
             var csName = $"cs-to-{dst.Database}";
             var result = await src.Maintenance.SendAsync(new PutConnectionStringOperation<RavenConnectionString>(new RavenConnectionString
